feat: modified files with Prettier and Eslint
