+++ src/components/search-input/SearchInput.tsx
@@ -7,7 +7,7 @@ import {
   setSearchValueState,
   State,
 } from '../../store/state';
-import { searchButtonClick } from '../search-button/SearchButton';
+import { searchButtonClick } from '../search-button/searchButtonClick';
 
 type InputProps = {
   placeholder?: string;
